\name{Prices}
\alias{Prices}
\docType{data}
\title{
A dummy data set of prices.
}
\description{
A dummy data set of prices or a non-stationary random process.
}
\usage{data(Prices)}
\format{
  The format is:
 num [1:168] 1394 1366 1499 1452 1421 ...
}

\source{
A dummy data set generated by the author.
}

\examples{

plot(Prices, type="l")
}
\keyword{datasets}
